# specs/001-build-a-modern/contracts/openapi.1.0.yaml
openapi: 3.1.0
info:
  title: LokiForm API
  summary: LokiForm API - A comprehensive form builder service that enables users to create, manage, and publish dynamic forms with advanced features including conditional logic, multiple field types, real-time validation, and submission management.
  description: LokiForm is a modern, feature-rich form builder API built with Next.js and TypeScript in a Turborepo monorepo architecture. The API provides complete form lifecycle management including creation, editing, publishing, and submission handling. Key features include drag-and-drop form building, multiple field types (text, email, select, checkbox, date, number), conditional logic, form validation, user authentication, role-based access control, and comprehensive submission management with pagination support.
  version: '1.0'
...
paths:
  /auth/change-password:
    post:
      operationId: Auth_changePassword
      summary: Change user password
      description: Updates the authenticated user's password by providing the current password and a new password. Validates the current password for security before applying the new password. This endpoint requires a valid Bearer authentication token and is used for secure password updates.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
              required:
                - oldPassword
                - newPassword
      security:
        - BearerAuth: []
  /auth/forgot-password:
    post:
      operationId: Auth_forgotPassword
      summary: Initiate password reset process
      description: Initiates a password reset process by sending a secure reset token to the user's registered email address. This endpoint is used when users forget their password and need to reset it. The reset token is time-limited and can be used with the reset-password endpoint.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
  /auth/login:
    post:
      operationId: Auth_signIn
      summary: User authentication and login
      description: Authenticates a user with their email and password credentials. Upon successful authentication, returns the user's complete profile information including personal details, preferences, and authentication status. This endpoint is the primary entry point for user authentication in the LokiForm application.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
  /auth/profile:
    get:
      operationId: Auth_getProfile
      summary: Get authenticated user profile
      description: Retrieves the authenticated user's complete profile information including personal details, account settings, interface preferences, and role information. This endpoint requires a valid Bearer authentication token and is used to display user information in the application interface.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      security:
        - BearerAuth: []
    put:
      operationId: Auth_updateProfile
      summary: Update user profile information
      description: Updates the authenticated user's profile information including personal details (name, email), password, interface preferences (theme mode, language), and account settings. All fields are optional and only provided fields will be updated. This endpoint requires a valid Bearer authentication token.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                password:
                  type: string
                interfaceMode:
                  type: string
                  enum:
                    - light
                    - dark
                    - system
                interfaceLanguage:
                  type: string
                  enum:
                    - en-US
                    - vi-VN
      security:
        - BearerAuth: []
  /auth/refresh:
    post:
      operationId: Auth_refresh
      summary: Refresh authentication token
      description: Refreshes an expired authentication token using a valid refresh token. This endpoint maintains user sessions without requiring re-authentication, providing a seamless user experience. Returns a new access token and refresh token pair upon successful validation.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
  /auth/register:
    post:
      operationId: Auth_register
      summary: Create new user account
      description: Creates a new user account with the provided personal information and credentials. Validates email uniqueness and password requirements before account creation. Returns the authenticated user's complete profile information upon successful registration. This endpoint enables new users to join the LokiForm platform.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - firstName
                - lastName
                - email
                - password
  /auth/reset-password:
    post:
      operationId: Auth_resetPassword
      summary: Complete password reset with token
      description: Completes the password reset process using a valid reset token received via email and a new password. Validates the token's authenticity and expiration before updating the password. This endpoint is used after the user receives a reset token via the forgot-password endpoint.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                newPassword:
                  type: string
              required:
                - token
                - newPassword
...
components:
  schemas:
...
    InterfaceLanguage:
      type: string
      enum:
        - en-US
        - vi-VN
      description: Defines the available language options for the user interface localization.
    InterfaceMode:
      type: string
      enum:
        - light
        - dark
        - system
      description: Defines the available interface mode options for user interface theming preferences.
...
    Role:
      type: string
      enum:
        - user
        - admin
        - superadmin
      description: Defines the available user roles in the LokiForm system, determining access levels and permissions for different types of users.
...
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - interfaceMode
        - interfaceLanguage
        - isActive
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/uuid'
          description: Unique identifier for the user
          readOnly: true
        firstName:
          type: string
          minLength: 1
          maxLength: 255
          description: First name of the user
        lastName:
          type: string
          minLength: 1
          maxLength: 255
          description: Last name of the user
        email:
          type: string
          minLength: 1
          maxLength: 255
          description: Email address of the user, used for authentication and communication
        role:
          allOf:
            - $ref: '#/components/schemas/Role'
          readOnly: true
        interfaceMode:
          allOf:
            - $ref: '#/components/schemas/InterfaceMode'
          description: User's preferred interface theme mode
        interfaceLanguage:
          allOf:
            - $ref: '#/components/schemas/InterfaceLanguage'
          description: User's preferred interface language
        isActive:
          type: boolean
          description: Whether the user account is active and can access the system
      description: Represents a user account in the LokiForm system with authentication credentials and profile information.
...
    uuid:
      type: string
      format: uuid
      description: Representing a universally unique identifier (UUID).
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:4000
    description: Development API endpoint
    variables: {}